refactoring bin to more modular
