+++ conf/modules.config
@@ -1693,48 +1693,25 @@ process {
     //
     // ANNOTATION
     //
-    withName: CUSTOM_MULTIQC_TABLES {
+    withName: CUSTOM_MULTIQC {
         ext.args   = [
                 params.save_intermediate_polishing ? "--save_intermediate" : '',
                 params.contig_filter_level ? "--filter_level ${params.contig_filter_level}" : '',
             ].join(' ').trim()
         publishDir = [
-            path: { "${params.outdir}/multiqc/custom_tables" },
-            mode: params.publish_dir_mode,
-            pattern: '*.tsv',
-            saveAs: { filename -> params.prefix || params.global_prefix  ? "${params.global_prefix}-$filename" : filename }
+            [
+                path: { "${params.outdir}/multiqc/custom_tables" },
+                mode: params.publish_dir_mode,
+                pattern: '*.tsv',
+                saveAs: { filename -> params.prefix || params.global_prefix  ? "${params.global_prefix}-$filename" : filename }
+            ],
+            [
+                path: { "${params.outdir}/multiqc" },
+                mode: params.publish_dir_mode,
+                pattern: 'multiqc*',
+                saveAs: { filename -> params.prefix || params.global_prefix  ? "${params.global_prefix}-$filename" : filename }
+            ]
         ]
     }
 
-    withName: MULTIQC_DATAPREP {
-        ext.args   = [
-            '-k tsv', // don't change outfmt as it is used by create_multiqc_custom_tables.py
-            '--no-report'
-        ].join(' ').trim()
-        publishDir = [
-            path: { "${params.outdir}/multiqc/multiqc_dataprep" },
-            mode: params.publish_dir_mode,
-            enabled: false,
-            pattern: 'multiqc*',
-            saveAs: { filename ->
-                filename.equals('versions.yml') ? null :
-                params.prefix || params.global_prefix  ? "${params.global_prefix}-$filename" : filename }
-        ]
-    }
-
-    withName: MULTIQC_REPORT {
-        ext.args   = [
-            params.multiqc_title ? "--title \"$params.multiqc_title\"" : '' ,
-            '-k tsv',
-            ].join(' ').trim()
-        publishDir = [
-            path: { "${params.outdir}/multiqc" },
-            mode: params.publish_dir_mode,
-            enabled: params.save_intermediate_polishing,
-            pattern: 'multiqc*',
-            saveAs: { filename ->
-                filename.equals('versions.yml') ? null :
-                params.prefix || params.global_prefix  ? "${params.global_prefix}-$filename" : filename }
-        ]
-    }
 }
